Refactor game event structures and state management in coinflip program
- Updated game-related events to use `u32` for `game_id` instead of `u64`, improving memory efficiency and consistency across the program.
- Simplified event emissions in `initialize_game`, `claim_win`, and `set_oracle_hash` instructions by removing unnecessary fields, enhancing clarity in event handling.
- Adjusted the `games_counter` and `id` fields in the `Oracle` and `Game` structs to use `u32`, aligning with the new event structure and improving data integrity.

These changes enhance the maintainability and clarity of game state and event management within the coinflip program.

+++ programs/coinflip/src/state.rs
@@ -9,7 +9,7 @@ pub struct Oracle {
     pub max_players: u16,
     pub max_timeout: i64,
     pub min_timeout: i64,
-    pub games_counter: u64,
+    pub games_counter: u32,
 }
 
 #[account]
@@ -50,7 +50,7 @@ impl Default for GameStatus {
 #[account]
 #[derive(Default)]
 pub struct Game {
-    pub id: u64,
+    pub id: u32,
     pub creator: Pubkey,
     pub game_type: GameType,
     pub amount: u64,
